ref: move corrupt messages fuzz tests
